pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plantinhas-auth-lib"
version = "0.1.0"
description = "Django authentication library for Plantinhas projects"
readme = "README.md"
authors = [{ name = "Marco Haubrich", email = "marcolino.haubrich@gmail.com" }]
maintainers = [
    { name = "Marco Haubrich", email = "marcolino.haubrich@gmail.com" },
]
keywords = ["django", "authentication", "jwt", "plantinhas"]
requires-python = ">=3.13.2"
dependencies = [
    "django-environ==0.12.0",
    "django>=5.2.1",
    "djangorestframework-simplejwt>=5.5.0",
    "pyjwt>=2.9.0",
    "requests>=2.32.3",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories
exclude = [
    ".git",
    ".ruff_cache",
    "__pypackages__",
    ".venv",
    "venv",
    "env",
    ".env",
    ".eggs",
    "*.egg",
    "dist",
    "build",
    "docs",
]

# Same as Black
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Skip magic trailing commas
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
known-first-party = ["plantinhas_auth_lib"]
known-third-party = ["django", "rest_framework"]
